+++ backend/src/main/java/com/ssafy/patpat/volunteer/dto/NoticeDto.java
@@ -13,7 +13,7 @@ import java.util.List;
 @ToString
 public class NoticeDto {
     private Long noticeId;
-    private Integer shelterId;
+    private Long shelterId;
     private String title;
     private List<VolunteerScheduleDto> list;
 }
